+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - CVE-2022-1049: Pcs daemon was allowing expired accounts, and accounts with
   expired passwords to login when using PAM auth. ([huntr#220307],
   [rhbz#2068456])
+- Pcsd does not expose the server name in HTTP headers anymore ([rhbz#2058278])
 
 ### Added
 - Add support for fence\_mpath to `pcs stonith update-scsi-devices` command
@@ -45,6 +46,7 @@
 [rhbz#1954099]: https://bugzilla.redhat.com/show_bug.cgi?id=1954099
 [rhbz#2023845]: https://bugzilla.redhat.com/show_bug.cgi?id=2023845
 [rhbz#2050274]: https://bugzilla.redhat.com/show_bug.cgi?id=2050274
+[rhbz#2058278]: https://bugzilla.redhat.com/show_bug.cgi?id=2058278
 [rhbz#2068456]: https://bugzilla.redhat.com/show_bug.cgi?id=2068456
 [huntr#220307]: https://huntr.dev/bounties/7aa921fc-a568-4fd8-96f4-7cd826246aa5/